requirements.txt
pdfplumber==0.11.9
rapidfuzz==3.14.1
pypdfium2==5.3.0
Pillow==12.1.0
pytesseract==0.3.13
ultralytics==8.3.2
